packages/scimarkdown/pyproject.toml
[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[project]
name = "scimarkdown"
version = "0.1.0+mit0.1.1"
description = "Scientific markdown converter with LaTeX formula detection and image extraction"
requires-python = ">=3.10"
dependencies = [
    "markitdown[all]>=0.1.1,<0.2.0",
    "Pillow>=9.0.0",
    "PyMuPDF>=1.24.0",
    "python-docx",
    "PyYAML",
]

[project.optional-dependencies]
ocr = ["pix2tex>=0.1.1"]
nougat = ["nougat-ocr"]
llm = ["openai>=1.0.0", "anthropic"]
embeddings = ["google-genai>=1.0.0"]
all = ["scimarkdown[ocr]", "scimarkdown[nougat]", "scimarkdown[llm]", "scimarkdown[embeddings]"]

[project.scripts]
scimarkdown = "scimarkdown.__main__:main"
scimarkdown-mcp = "scimarkdown.mcp.__main__:main"
